레시피 저장 테스트 refactor
엔티티에 updatable=false를 제거했습니다. 레피시 저장 테스트를 위해 테스트 파일 추가했습니다. RecipeCreateDto의 Valid를 조정하고 그로 인해 사용 하지 않는 ErrorMessage를 제거했습니다. 차후 레시피 관련 작업을 위해 FailCode에 레시피 관련 Code를 정의했습니다.

+++ src/main/java/com/my/bob/core/domain/recipe/service/RecipeSaveService.java
@@ -3,5 +3,5 @@ package com.my.bob.core.domain.recipe.service;
 import com.my.bob.core.domain.recipe.dto.request.RecipeCreateDto;
 
 public interface RecipeSaveService {
-    long newRecipe(RecipeCreateDto dto);
+    int newRecipe(RecipeCreateDto dto);
 }
